+++ glade_forms/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.2.0
+- **[Add]**: Add `resetToPure` on model level.
+
+## 2.1.0
+- **[Add]**: Add `defaultTranslateError` on model level.
+
 ## 2.0.1
 - **[Fix]**: Fix `isUri()` to handle URL corectly